requirements.txt
streamlit
pandas
pdfplumber>=0.10.0
python-docx